LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontfaceset-load-css-wide-keywords-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: SyntaxError thrown when matching CSS-wide keyword
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

PASS Loading CSS-wide keyword "initial" causes SyntaxError (document)
PASS Loading CSS-wide keyword "inherit" causes SyntaxError (document)
PASS Loading CSS-wide keyword "unset" causes SyntaxError (document)
PASS Loading CSS-wide keyword "revert" causes SyntaxError (document)
PASS Loading CSS-wide keyword "initial" causes SyntaxError (worker)
PASS Loading CSS-wide keyword "inherit" causes SyntaxError (worker)
PASS Loading CSS-wide keyword "unset" causes SyntaxError (worker)
PASS Loading CSS-wide keyword "revert" causes SyntaxError (worker)
